Add debos recipe for raspberrypi ARM64-based boards
This adds a debos recipe for building a desktop image for a ARM64-based
RaspberryPi. It include a lightweight desktop, that is LXDE, create a
default user "pi" that has sudo access rights. As on Raspbian, the "pi"
user is automatically logged in via lightdm. The buster debian splash is
also automatically enabled